requirements.txt
pika~=1.3.2
pillow~=10.0.1
rioxarray~=0.15.0
rasterio~=1.3.8
gdal2tiles~=0.1.9
numpy~=1.26.0
python-dotenv~=1.0.0
pydantic~=2.4.2
pyproj~=3.6.0